Refactored dnnhmm class to use no more hmmlearn, manually implemented general-purpose viterbi recognizer.
